% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{confus}
\alias{confus}
\title{Inward facing function}
\usage{
confus(rst, landcover)
}
\arguments{
\item{rst}{a raster of either soil aglim or slope}

\item{landcover}{a categorized landcover map}
}
\description{
Inward facing function
}
