% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_farmers.R
\name{distrubution_by_farmer}
\alias{distrubution_by_farmer}
\title{A function to fields between farmers on a LGrafEU output object}
\usage{
distrubution_by_farmer(
  output_obj,
  assign_mode = 1,
  mean_fields_per_farm,
  sd_fields_per_farm,
  distribution
)
}
\arguments{
\item{output_obj}{a LGrafEU output object}

\item{assign_mode}{1 = unstructured distribution 2 = spatially strcutured distribution of fields}

\item{mean_fields_per_farm}{mean fields per farmer}

\item{sd_fields_per_farm}{sd fields per farmer}

\item{distribution}{either norm or lnorm}
}
\value{
A LGraf output object with fields assined to farmers
}
\description{
A function to fields between farmers on a LGrafEU output object
}
\examples{
r<-terra::rast(matrix(1, nrow=50, ncol=50))
map<-establish_by_place_conquer(potential_space= r,
                                cell_size=1,
                                includsion_value = 1,
                                mean_field_size = 100,
                                sd_field_size = 25,
                                distribution = "norm",
                                mean_shape_index = 1,
                                sd_shape_index = 0.3,
                                percent = 75,
                                assign_farmers = TRUE,
                                assign_mode = 2,
                                mean_fields_per_farm = 3,
                                sd_fields_per_farm = 3)


outcome<-distrubution_by_farmer(map,assign_mode = 1, mean_fields_per_farm =10, sd_fields_per_farm = 5, distribution = "norm")

return_by_farmer(outcome)
}
