% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_crops.R
\name{distrubution_by_percent}
\alias{distrubution_by_percent}
\title{Title}
\usage{
distrubution_by_percent(output_obj, crops_matrix)
}
\arguments{
\item{crops_matrix}{}
}
\description{
Title
}
\examples{
r<-raster::raster(matrix(1, nrow=50, ncol=50))
map<-establish_by_place_conquer(potential_space= r,
                                cell_size=1,
                                includsion_value = 1,
                                mean_field_size = 100,
                                sd_field_size = 25,
                                distribution = "norm",
                                mean_shape_index = 1,
                                sd_shape_index = 0.3,
                                percent = 75,
                                assign_farmers = TRUE,
                                assign_mode = 2,
                                mean_fields_per_farm = 3,
                                sd_fields_per_farm = 3)


crops_matrix <- data.frame(crop = c("Wheat", "Corn", "Soybean"),
                          percentage = c(0.4, 0.3, 0.3),
                          index = c(1,2,3))  # Desired percentages

outcome<-distrubution_by_percent(map,crops_matrix)

plot_by_crop(outcome)
}
