% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_by_dead_leaves.R
\name{establish_by_dead_leaves}
\alias{establish_by_dead_leaves}
\title{Field establishment based on a dead leave algorithm}
\usage{
establish_by_dead_leaves(
  potential_space,
  cell_size,
  includsion_value,
  mean_field_size,
  sd_field_size,
  mean_shape_index,
  sd_shape_index,
  percent,
  assign_farmers,
  assign_mode,
  mean_fields_per_farm,
  sd_fields_per_farm
)
}
\arguments{
\item{sd_fields_per_farm}{}
}
\description{
Field establishment based on a dead leave algorithm
}
\examples{
r<-raster::raster(matrix(1, nrow=50, ncol=50))
raster::extent(r)<-c(0,100,0,100)
dead_leaves_texture <- establish_by_dead_leaves(potential_space = r,
                                                cell_size = 1,
                                                includsion_value = 1,
                                                mean_field_size = 50,
                                                sd_field_size = 25,
                                                mean_shape_index = .5,
                                                sd_shape_index = .1,
                                                percent = 0.75,
                                                assign_farmers = TRUE,
                                                assign_mode = 2,
                                                mean_fields_per_farm = 4,
                                                sd_fields_per_farm = 4)
raster::plot(dead_leaves_texture$map)
}
