% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_by_space.R
\name{establish_by_space}
\alias{establish_by_space}
\title{Title}
\usage{
establish_by_space(
  potential_space,
  mean_field_size,
  sd_field_size,
  mean_shape_index,
  sd_shape_index,
  percent,
  assign_farmers,
  mean_fields_per_farm
)
}
\arguments{
\item{potential_space}{a map containing potential space for field establishment}

\item{mean_field_size}{mean field size}

\item{sd_field_size}{sd field size}

\item{mean_shape_index}{mean shape index}

\item{sd_shape_index}{sd shape index}

\item{percent}{percent of the potential arabale land that will be filled ranging from 0-100}

\item{assign_farmers}{TRUE/FALSE to assign fields to random farmers}

\item{mean_fields_per_farm}{mean number of fields owned by each farmer}
}
\description{
Title
}
