% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_by_dead_leaves.R
\name{generate_dead_leaves}
\alias{generate_dead_leaves}
\title{Field establishment based on a dead leave algorithm}
\usage{
generate_dead_leaves(
  potential_space,
  cell_size,
  includsion_value,
  mean_field_size,
  sd_field_size,
  mean_shape_index,
  sd_shape_index,
  percent,
  assign_farmers,
  assign_mode,
  mean_fields_per_farm,
  sd_fields_per_farm
)
}
\arguments{
\item{sd_fields_per_farm}{}
}
\description{
Field establishment based on a dead leave algorithm
}
\examples{
r<-raster(matrix(1, nrow=50, ncol=50))
extent(r)<-c(0,100,0,100)
dead_leaves_texture <- generate_dead_leaves(r,1,1,50,25,.5,.1, 0.75, TRUE,2, 4,4)
plot(dead_leaves_texture$map)
}
