% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranlan.R
\name{generate_potential_clusters}
\alias{generate_potential_clusters}
\title{A natural agriculture cluster map for creating artificial potential agricultural landscape}
\usage{
generate_potential_clusters(rows, cols, p, ai)
}
\arguments{
\item{rows}{nrows of raster}

\item{cols}{ncols of raster}

\item{p}{proportion of element that forms a clustre}

\item{ai}{percentage of each category of landscape}
}
\description{
This is a wramper around the nlmr random cluster function to create naturalist potential agriculture space raster
}
