% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranlan.R
\name{generate_potential_landscape}
\alias{generate_potential_landscape}
\title{A natural agriculture cluster map for creating artificial potential agricultural landscape}
\usage{
generate_potential_landscape(rows, cols, p, ai, rescale)
}
\arguments{
\item{p}{f}

\item{ai}{percentage of each category of landscape}

\item{rescale}{}
}
\description{
This is a wramper around the nlmr random cluster function to create naturalist potential agriculture space raster
}
