% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roads_least_cost.R
\name{generate_route_lc}
\alias{generate_route_lc}
\title{A function for generating artificial roads}
\usage{
generate_route_lc(road_length, slope)
}
\arguments{
\item{road_length}{A number}

\item{slope}{A lope raster}
}
\value{
A raster
}
\description{
A function for generating artificial roads
}
\examples{
test<-generate_perlin_noise(200,200,1,2,3,0.005,FALSE, "land_percentage", percetange = 50)
test2<-generate_route_lc(2000,test)
plot(test2)

}
