% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranlan.R
\name{generate_slope}
\alias{generate_slope}
\title{A perlin noise based slope generator}
\usage{
generate_slope(
  width,
  height,
  cellSize,
  frequency,
  octaves,
  lacunarity,
  categorized
)
}
\arguments{
\item{categorized}{TRUE/FLASE tells you if the slope raster returns categorized or smooth}
}
\description{
A perlin noise based slope generator
}
