% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LGrafEU-package.R
\docType{package}
\name{LGrafEU-package}
\alias{LGrafEU}
\alias{LGrafEU-package}
\title{LGrafEU: Agricultural landscape generator}
\description{
Generates artificial agricultural landscapes <link to publicationn> Artificial agricultural landscape generators are designed to systematically and realistically produce artificial maps of agricutltural dominated landscapes, which are meant to be used as an input for ecological models.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/pogoyoly/LGrafEU/issues}
}

}
\author{
\strong{Maintainer}: Eyal Goldstein \email{pogoyoly@gmai.com}

Other contributors:
\itemize{
  \item Antonia Deutcher \email{antdeutscher@gmail.com} [contributor]
  \item Eamon O'keeffe \email{eamon.okeeffe@stud.uni-goettingen.de} [contributor]
  \item Sebastian Hanss \email{shanss@uni-goettingen.de} [contributor]
}

}
\keyword{internal}
