% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_by_village.R
\name{placerandomvillage}
\alias{placerandomvillage}
\title{Title}
\usage{
placerandomvillage(road_raster, num_villages)
}
\arguments{
\item{road_raster}{A raster generated by one of the road algorithms}

\item{num_villages}{An integer for number of villages that are going to be placed}
}
\description{
Title
}
