% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_by_crop}
\alias{plot_by_crop}
\title{Plot by field number}
\usage{
plot_by_crop(output_obj)
}
\arguments{
\item{output_obj}{an output object of one of the establish functions}
}
\description{
Plot by field number
}
