% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_by_field}
\alias{plot_by_field}
\title{Plot by field number}
\usage{
plot_by_field(output_obj, method = 1)
}
\arguments{
\item{output_obj}{an output object of one of the establish functions}

\item{method}{}
}
\description{
Plot by field number
}
