% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{return_by_crop}
\alias{return_by_crop}
\title{Plot by crop type}
\usage{
return_by_crop(output_obj, method = 1)
}
\arguments{
\item{output_obj}{an output object of one of the establish functions}

\item{method}{1 is plot directly 2 returns a raster of crop type}
}
\value{
method = 1 plots a raster and method = 2 returns an LGrafEU output object
}
\description{
Plot by crop type
}
