% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{return_by_farmer}
\alias{return_by_farmer}
\title{return by farmer}
\usage{
return_by_farmer(output_obj, method = 1)
}
\arguments{
\item{output_obj}{and output object of one of the establish functions}

\item{method}{1 is plot directly 2 returns a raster of crop type}
}
\description{
return by farmer
}
