% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{return_by_field}
\alias{return_by_field}
\title{Plot by field number}
\usage{
return_by_field(output_obj, method = 1)
}
\arguments{
\item{output_obj}{an output object of one of the establish functions}

\item{method}{1 is plot directly 2 returns a raster of crop type}
}
\description{
Plot by field number
}
