% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{sl_txt_ag}
\alias{sl_txt_ag}
\title{Generate using artificial rasters}
\usage{
sl_txt_ag(
  texture_gen,
  slope_gen,
  aglim_gen,
  aglim_real,
  slope_real,
  texture_real,
  landcover
)
}
\arguments{
\item{texture_gen}{artificially generated texture raster}

\item{slope_gen}{artificially generated lope raster}

\item{aglim_gen}{artificially generated aglim raster}

\item{aglim_real}{location aglim raster}

\item{slope_real}{location lope raster}

\item{texture_real}{location texture raster}

\item{landcover}{location landcover raster}
}
\value{
An artificially generated landcover file
}
\description{
Convert artificial rasters into artificial landcover
}
