% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{trans}
\alias{trans}
\title{Title}
\usage{
trans(transition, rst)
}
\arguments{
\item{transition}{the transition matrix from the confus function}

\item{rst}{the raster from which the transition occurs}
}
\description{
Title
}
