% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_by_dead_leaves.R
\name{establish_by_dead_leaves}
\alias{establish_by_dead_leaves}
\title{Field establishment based on a dead leave algorithm}
\usage{
establish_by_dead_leaves(
  potential_space,
  cell_size,
  includsion_value,
  mean_field_size,
  sd_field_size,
  distribution = "norm",
  mean_shape_index,
  sd_shape_index,
  percent,
  assign_farmers,
  assign_mode,
  mean_fields_per_farm,
  sd_fields_per_farm
)
}
\arguments{
\item{potential_space}{a raster including a potential space category for field placement}

\item{cell_size}{cell size for output}

\item{includsion_value}{inclusion value for the potential space raster}

\item{mean_field_size}{mean field size counted by number of cells from a normal distribution}

\item{sd_field_size}{sd field size counted by number of cells from a normal distribution}

\item{distribution}{distribution type for field size}

\item{mean_shape_index}{mean shape index calculated by a relation between width/length of placement}

\item{sd_shape_index}{sd shape index calculated by a relation between width/length of placement}

\item{percent}{percent of the potential space to be filled with fields}

\item{assign_farmers}{TRUE/FALSE for farmer assignment in model}

\item{assign_mode}{1 = random assignment 2 = spatially structured assignment}

\item{mean_fields_per_farm}{mean fields per farmer from a long normal distribution}

\item{sd_fields_per_farm}{fields per farmer from a long normal distribution}
}
\value{
LGraf object
}
\description{
Field establishment based on a dead leave algorithm
}
\examples{
r<-raster::raster(matrix(1, nrow=50, ncol=50))
raster::extent(r)<-c(0,100,0,100)
output <- establish_by_dead_leaves(potential_space = r,
                                                cell_size = 1,
                                                includsion_value = 1,
                                                mean_field_size = 50,
                                                sd_field_size = 25,
                                                distribution = "norm",
                                                mean_shape_index = .5,
                                                sd_shape_index = .1,
                                                percent = 0.75,
                                                assign_farmers = TRUE,
                                                assign_mode = 2,
                                                mean_fields_per_farm = 4,
                                                sd_fields_per_farm = 4)
plot_by_field(output)
}
