% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_by_space.R
\name{establish_by_space}
\alias{establish_by_space}
\title{Heuristic field establishment based on simple geometries}
\usage{
establish_by_space(
  potential_space,
  cell_size,
  includsion_value,
  additional_lim = NA,
  mean_field_size,
  sd_field_size,
  mean_shape_index,
  sd_shape_index,
  percent,
  assign_farmers,
  assign_mode,
  mean_fields_per_farm,
  sd_fields_per_farm
)
}
\arguments{
\item{sd_fields_per_farm}{}
}
\description{
Heuristic field establishment based on simple geometries
}
