% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_by_dead_leaves.R
\name{generate_dead_leaves}
\alias{generate_dead_leaves}
\title{Title}
\usage{
generate_dead_leaves(
  potential_space,
  cell_size,
  includsion_value,
  mean_field_size,
  sd_field_size,
  mean_shape_index,
  sd_shape_index,
  percent,
  assign_farmers,
  assign_mode,
  mean_fields_per_farm,
  sd_fields_per_farm
)
}
\arguments{
\item{sd_fields_per_farm}{}
}
\description{
Title
}
\examples{
r <- setValues(raster(nrows=200, ncols=200), 1)
dead_leaves_texture <- generate_dead_leaves(r,1,1,25,10,.5,.1, 0.75, TRUE,2, 4,4)
plot(dead_leaves_texture$map)
}
