% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranlan.R
\name{generate_soil}
\alias{generate_soil}
\title{Natural soil cluster generator for artificial aglim and soil texture maps}
\usage{
generate_soil(rows, cols, p, ai, rescale)
}
\arguments{
\item{rescale}{}
}
\description{
Natural soil cluster generator for artificial aglim and soil texture maps
}
\details{
This is a wramper around the nlmr random cluster function to create naturalist soil raster
}
