% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_by_arable_land}
\alias{plot_by_arable_land}
\title{Plot by arable land}
\usage{
plot_by_arable_land(output_obj)
}
\arguments{
\item{output_obj}{an output object of one of the establish functions}
}
\description{
Plot by arable land
}
