% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_by_farmer}
\alias{plot_by_farmer}
\title{Plot by farmer}
\usage{
plot_by_farmer(output_obj)
}
\arguments{
\item{output_obj}{and output object of one of the establish functions}
}
\description{
Plot by farmer
}
