% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{trans_1lr}
\alias{trans_1lr}
\title{A one layer transformation}
\usage{
trans_1lr(rast, landcover, aggregation)
}
\arguments{
\item{rast}{A type feature raster}

\item{landcover}{A landcover raster}

\item{aggregation}{A number that defines how aggregated he potential space will be}
}
\value{
A raster transformed
}
\description{
A one layer transformation
}
