% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chains.R
\name{trans_3lr}
\alias{trans_3lr}
\title{A three layer transformation}
\usage{
trans_3lr(texture, slope, aglim)
}
\arguments{
\item{texture}{A soil texture raster}

\item{slope}{A categorized slope raster}

\item{aglim}{A agricultural limitation raster}
}
\description{
A three layer transformation
}
